requirements.txt
playwright==1.41.0
apscheduler==3.10.4
openai==1.12.0
anthropic==0.18.0
python-dotenv==1.0.1
httpx<0.28
playwright-stealth==2.0.3
browserforge==1.2.4
PyQt5==5.15.10
